dashboard/serializers.py
from rest_framework_gis import serializers
from .models import Marker

class MarkerSerializer(serializers.GeoFeatureModelSerializer):
    class Meta:
        model = Marker
        geo_field = "location"
        fields = ("id", "name")